+++ lib/database/lib/model/remote-clone-pluggable-database-details.ts
@@ -16,6 +16,7 @@ import * as model from "../model";
 import common = require("oci-common");
 
 /**
+ * **Deprecated.** Use {@link #createPluggableDatabaseDetails(CreatePluggableDatabaseDetailsRequest) createPluggableDatabaseDetails} for Pluggable Database RemoteClone Operation.
  * Parameters for cloning a pluggable database (PDB) in a remote database (CDB). A remote CDB is one that does not contain the source PDB.
  * <p>
  **Warning:** Oracle recommends that you avoid using any confidential information when you supply string values using the API.