+++ lib/marketplace/lib/marketplace-waiter.ts
@@ -41,4 +41,20 @@ export class MarketplaceWaiter {
       targetStates.includes(models.Publication.LifecycleState.Deleted)
     );
   }
+
+  /**
+   * Waits forWorkRequest
+   *
+   * @param request the request to send
+   * @return response returns GetWorkRequestResponse
+   */
+  public async forWorkRequest(
+    request: serviceRequests.GetWorkRequestRequest
+  ): Promise<serviceResponses.GetWorkRequestResponse> {
+    return genericWaiter(
+      this.config,
+      () => this.client.getWorkRequest(request),
+      response => (response.workRequest.timeFinished ? true : false)
+    );
+  }
 }