+++ lib/vnmonitoring/lib/model/capture-filter.ts
@@ -70,6 +70,11 @@ Example: `2021-08-25T21:10:29.600Z`
    *
    */
   "vtapCaptureFilterRules"?: Array<model.VtapCaptureFilterRuleDetails>;
+  /**
+   * The set of rules governing what traffic the Flow Log collects when creating a flow log capture filter.
+   *
+   */
+  "flowLogCaptureFilterRules"?: Array<model.FlowLogCaptureFilterRuleDetails>;
 }
 
 export namespace CaptureFilter {
@@ -88,6 +93,7 @@ export namespace CaptureFilter {
 
   export enum FilterType {
     Vtap = "VTAP",
+    Flowlog = "FLOWLOG",
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this
      * version of the SDK.
@@ -103,6 +109,11 @@ export namespace CaptureFilter {
           ? obj.vtapCaptureFilterRules.map(item => {
               return model.VtapCaptureFilterRuleDetails.getJsonObj(item);
             })
+          : undefined,
+        "flowLogCaptureFilterRules": obj.flowLogCaptureFilterRules
+          ? obj.flowLogCaptureFilterRules.map(item => {
+              return model.FlowLogCaptureFilterRuleDetails.getJsonObj(item);
+            })
           : undefined
       }
     };
@@ -116,6 +127,11 @@ export namespace CaptureFilter {
         "vtapCaptureFilterRules": obj.vtapCaptureFilterRules
           ? obj.vtapCaptureFilterRules.map(item => {
               return model.VtapCaptureFilterRuleDetails.getDeserializedJsonObj(item);
+            })
+          : undefined,
+        "flowLogCaptureFilterRules": obj.flowLogCaptureFilterRules
+          ? obj.flowLogCaptureFilterRules.map(item => {
+              return model.FlowLogCaptureFilterRuleDetails.getDeserializedJsonObj(item);
             })
           : undefined
       }