+++ lib/vnmonitoring/lib/model/create-capture-filter-details.ts
@@ -53,11 +53,17 @@ export interface CreateCaptureFilterDetails {
    *
    */
   "vtapCaptureFilterRules"?: Array<model.VtapCaptureFilterRuleDetails>;
+  /**
+   * The set of rules governing what traffic the Flow Log collects when creating a flow log capture filter.
+   *
+   */
+  "flowLogCaptureFilterRules"?: Array<model.FlowLogCaptureFilterRuleDetails>;
 }
 
 export namespace CreateCaptureFilterDetails {
   export enum FilterType {
-    Vtap = "VTAP"
+    Vtap = "VTAP",
+    Flowlog = "FLOWLOG"
   }
 
   export function getJsonObj(obj: CreateCaptureFilterDetails): object {
@@ -67,6 +73,11 @@ export namespace CreateCaptureFilterDetails {
         "vtapCaptureFilterRules": obj.vtapCaptureFilterRules
           ? obj.vtapCaptureFilterRules.map(item => {
               return model.VtapCaptureFilterRuleDetails.getJsonObj(item);
+            })
+          : undefined,
+        "flowLogCaptureFilterRules": obj.flowLogCaptureFilterRules
+          ? obj.flowLogCaptureFilterRules.map(item => {
+              return model.FlowLogCaptureFilterRuleDetails.getJsonObj(item);
             })
           : undefined
       }
@@ -81,6 +92,11 @@ export namespace CreateCaptureFilterDetails {
         "vtapCaptureFilterRules": obj.vtapCaptureFilterRules
           ? obj.vtapCaptureFilterRules.map(item => {
               return model.VtapCaptureFilterRuleDetails.getDeserializedJsonObj(item);
+            })
+          : undefined,
+        "flowLogCaptureFilterRules": obj.flowLogCaptureFilterRules
+          ? obj.flowLogCaptureFilterRules.map(item => {
+              return model.FlowLogCaptureFilterRuleDetails.getDeserializedJsonObj(item);
             })
           : undefined
       }